feat: Order API 권한별 접근 개발 완료

+++ src/main/java/com/f1/fastone/order/service/OrderService.java
@@ -3,8 +3,10 @@ package com.f1.fastone.order.service;
 import com.f1.fastone.cart.entity.Cart;
 import com.f1.fastone.cart.entity.CartItem;
 import com.f1.fastone.cart.repository.CartRepository;
+import com.f1.fastone.common.auth.security.UserDetailsImpl;
 import com.f1.fastone.common.exception.ErrorCode;
 import com.f1.fastone.common.exception.custom.EntityNotFoundException;
+import com.f1.fastone.common.exception.custom.ServiceException;
 import com.f1.fastone.menu.entity.Menu;
 import com.f1.fastone.order.dto.OrderItemDto;
 import com.f1.fastone.order.dto.PaymentDto;
@@ -21,15 +23,19 @@ import com.f1.fastone.store.entity.Store;
 import com.f1.fastone.order.dto.request.OrderRequestDto;
 import com.f1.fastone.store.repository.StoreRepository;
 import com.f1.fastone.user.entity.User;
+import com.f1.fastone.user.entity.UserRole;
 import com.f1.fastone.user.repository.UserRepository;
 import jakarta.transaction.Transactional;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 
+import java.time.Duration;
+import java.time.LocalDateTime;
 import java.util.List;
 import java.util.UUID;
 
 import lombok.extern.slf4j.Slf4j;
+
 
 @Slf4j
 @Service
@@ -44,10 +50,11 @@ public class OrderService {
 
     @Transactional
     public OrderResponseDto createOrder(String username, OrderRequestDto requestDto) {
-        // User, 가게, 장바구니 조회
+        // User(CUSTOMER), 가게, 장바구니 조회
         User user = userRepository.findByUsername(username).orElseThrow(() -> new EntityNotFoundException(ErrorCode.USER_NOT_FOUND));
         Store store = storeRepository.findById(requestDto.getStoreId()).orElseThrow(() -> new EntityNotFoundException(ErrorCode.STORE_NOT_FOUND));
         Cart cart = cartRepository.findById(requestDto.getCartId()).orElseThrow(() -> new EntityNotFoundException(ErrorCode.CART_NOT_FOUND));
+
         // Order 생성
         Order order = requestDto.toEntity(user, store);
         orderRepository.save(order);
@@ -60,24 +67,31 @@ public class OrderService {
         order.setOrderItems(orderItems);
 
         PaymentDto paymentDto = new PaymentDto(order.getTotalPrice());
-
-        return new OrderResponseDto(order.getCreatedAt(), order.getStore().getName(), orderItemDtos, paymentDto, order.getStatus());
+        return new OrderResponseDto(order.getId(), order.getCreatedAt(), order.getStore().getName(), orderItemDtos, paymentDto, order.getStatus());
     }
 
     @Transactional
     public List<OrderResponseDto> getOrders(String username) {
         User user = userRepository.findByUsername(username).orElseThrow(() -> new EntityNotFoundException(ErrorCode.USER_NOT_FOUND));
-        List<Order> orders = user.getOrders(); // null;
+        UserRole role = user.getRole();
+        List<Order> orders = null;
 
-//        switch (username) {
-//            case "MANAGER" -> orders = orderRepository.findAll();
-//            case "OWNER" -> {
-//                // 가게별로 Order에 있는 가게 주인과 Username 또는 UserDetails 대조해보고 조회하기
-//            }
-//            case "CUSTOMER" -> orders = user.getOrders();
-//            default -> {
-//            }
-//        }
+        switch (role) {
+            case CUSTOMER -> {
+                orders = user.getOrders();
+            }
+            case OWNER -> {
+                Store store = storeRepository.findByOwner(user);
+                orders = orderRepository.findAllByStore(store);
+            }
+            case MANAGER -> {
+                orders = orderRepository.findAll();
+            }
+        }
+
+        if (orders == null || orders.isEmpty()) {
+            throw new EntityNotFoundException(ErrorCode.ORDER_NOT_FOUND);
+        }
 
         return orders.stream().map(order ->
                 OrderResponseDto.from(
@@ -89,8 +103,23 @@ public class OrderService {
     }
 
     @Transactional
-    public OrderDetailResponseDto getOrderDetail(UUID orderId) {
-        Order order = orderRepository.findById(orderId).orElseThrow(() -> new EntityNotFoundException(ErrorCode.ORDER_NOT_FOUND));
+    public OrderDetailResponseDto getOrderDetail(String username, UUID orderId) {
+        User user = userRepository.findByUsername(username).orElseThrow(() -> new EntityNotFoundException(ErrorCode.USER_NOT_FOUND));
+        UserRole role = user.getRole();
+        Order order = null;
+
+        switch (role) {
+            case CUSTOMER -> {
+                order = orderRepository.findByIdAndUser(orderId, user).orElseThrow(() -> new EntityNotFoundException(ErrorCode.ORDER_DETAIL_ACCESS_DENIED));
+            }
+            case OWNER -> {
+                Store store = storeRepository.findByOwner(user);
+                order = orderRepository.findByIdAndStore(orderId, store).orElseThrow(() -> new EntityNotFoundException(ErrorCode.ORDER_DETAIL_ACCESS_DENIED));;
+            }
+            case MANAGER -> {
+                order = orderRepository.findById(orderId).orElseThrow(() -> new EntityNotFoundException(ErrorCode.ORDER_DETAIL_ACCESS_DENIED));
+            }
+        }
 
         ShipToDto shipToDto = new ShipToDto(
             order.getShipToName(), order.getShipToPhone(), order.getPostalCode(), order.getCity(), order.getAddress(), order.getAddressDetail()
@@ -102,25 +131,57 @@ public class OrderService {
     }
 
     @Transactional
-    public OrderResponseDto updateOrderStatus(UUID orderId, OrderStatusRequestDto requestDto) {
-        OrderStatus status = requestDto.getOrderStatus();
+    public OrderResponseDto updateOrderStatus(UserDetailsImpl userDetails, UUID orderId, OrderStatusRequestDto requestDto) {
+        User user = userDetails.getUser();
+        UserRole role = user.getRole();
+        Order order = null;
 
-        Order order = orderRepository.findById(orderId).orElseThrow(() -> new EntityNotFoundException(ErrorCode.ORDER_NOT_FOUND));
+        switch (role) {
+            case CUSTOMER -> {
+                throw new ServiceException(ErrorCode.ORDER_UPDATE_DENIED);
+            }
+            case OWNER -> {
+                Store store = storeRepository.findByOwner(user);
+                order = orderRepository.findByIdAndStore(orderId, store).orElseThrow(() -> new ServiceException(ErrorCode.ORDER_UPDATE_DENIED));;
+            }
+            case MANAGER -> {
+                order = orderRepository.findById(orderId).orElseThrow(() -> new EntityNotFoundException(ErrorCode.ORDER_NOT_FOUND));
+            }
+        }
+
+        OrderStatus status = requestDto.getOrderStatus();
         order.setOrderStatus(status);
 
         List<OrderItemDto> orderItemDtos = order.getOrderItems().stream().map(OrderItemDto::from).toList();
 
         PaymentDto paymentDto = new PaymentDto(order.getTotalPrice());
 
-        return new OrderResponseDto(order.getCreatedAt(), order.getStore().getName(), orderItemDtos, paymentDto, order.getStatus());
+        return new OrderResponseDto(order.getId(), order.getCreatedAt(), order.getStore().getName(), orderItemDtos, paymentDto, order.getStatus());
     }
 
     @Transactional
-    public void deleteOrder(UUID orderId) {
+    public void deleteOrder(UserDetailsImpl userDetails, UUID orderId) {
+        User user = userDetails.getUser();
+        UserRole role = user.getRole();
+
         Order order = orderRepository.findById(orderId).orElseThrow(
                 () -> new EntityNotFoundException(ErrorCode.ORDER_NOT_FOUND)
         );
-        orderRepository.delete(order);
+
+        switch (role) {
+            case CUSTOMER -> {
+                LocalDateTime orderCreatedAt = order.getCreatedAt();
+                LocalDateTime now = LocalDateTime.now();
+                if (Duration.between(orderCreatedAt, now).toMinutes() <= 5) {
+                    orderRepository.delete(order);
+                } else {
+                    throw new ServiceException(ErrorCode.ORDER_DELETE_DENIED);
+                }
+            }
+            case OWNER, MANAGER -> {
+                orderRepository.delete(order);
+            }
+        }
     }
 
     private List<OrderItem> convertCartToOrder(Order order, List<CartItem> cartItems) {
